port AddressSorter class over to uk_geo_utils app
