+++ polling_stations/apps/data_finder/helpers.py
@@ -280,43 +280,6 @@ def get_council(geocode_result):
     return Council.objects.defer("area", "location").get(area__covers=location)
 
 
-class AddressSorter:
-    # Class for sorting sort a list of address objects
-    # in a human-readable order.
-
-    def __init__(self, addresses):
-        self.addresses = addresses
-
-    def convert(self, text):
-        # if text is numeric, covert to an int
-        # this allows us to sort numbers in int order, not string order
-        return int(text) if text.isdigit() else text
-
-    def alphanum_key(self, tup):
-        # split the desired component of tup (defined by key function)
-        # into a listof numeric and text components
-        return [ self.convert(c) for c in filter(None, re.split('([0-9]+)', tup[1])) ]
-
-    def swap_fields(self, item):
-        lst = self.alphanum_key(item)
-        # swap things about so we can sort by street name, house number
-        # instead of house number, street name
-        if len(lst) > 1 and isinstance(lst[0], int) and isinstance(lst[1], str) and (lst[1][0].isspace() or lst[1][0] == ','):
-            lst[0], lst[1] = lst[1], lst[0]
-        if len(lst) > 1 and isinstance(lst[0], int) and isinstance(lst[1], int):
-            lst[0], lst[1] = lst[1], lst[0]
-        if isinstance(lst[0], int):
-            lst[0] = str(lst[0])
-        return lst
-
-    def natural_sort(self):
-        sorted_list = sorted(
-            [(address, address.address) for address in self.addresses],
-            key=self.swap_fields
-        )
-        return [address[0] for address in sorted_list]
-
-
 class EveryElectionWrapper:
 
     def __init__(self, postcode):
