+++ polling_stations/apps/uk_geo_utils/helpers.py
@@ -125,3 +125,40 @@ class AddressFormatter:
     def __str__(self):
         """Return the label form of the address."""
         return ','.join(self.generate_address_label())
+
+
+class AddressSorter:
+    # Class for sorting sort a list of address objects
+    # in a human-readable order.
+
+    def __init__(self, addresses):
+        self.addresses = addresses
+
+    def convert(self, text):
+        # if text is numeric, covert to an int
+        # this allows us to sort numbers in int order, not string order
+        return int(text) if text.isdigit() else text
+
+    def alphanum_key(self, tup):
+        # split the desired component of tup (defined by key function)
+        # into a listof numeric and text components
+        return [ self.convert(c) for c in filter(None, re.split('([0-9]+)', tup[1])) ]
+
+    def swap_fields(self, item):
+        lst = self.alphanum_key(item)
+        # swap things about so we can sort by street name, house number
+        # instead of house number, street name
+        if len(lst) > 1 and isinstance(lst[0], int) and isinstance(lst[1], str) and (lst[1][0].isspace() or lst[1][0] == ','):
+            lst[0], lst[1] = lst[1], lst[0]
+        if len(lst) > 1 and isinstance(lst[0], int) and isinstance(lst[1], int):
+            lst[0], lst[1] = lst[1], lst[0]
+        if isinstance(lst[0], int):
+            lst[0] = str(lst[0])
+        return lst
+
+    def natural_sort(self):
+        sorted_list = sorted(
+            [(address, address.address) for address in self.addresses],
+            key=self.swap_fields
+        )
+        return [address[0] for address in sorted_list]
